Move out extract_extending_graph
